Refactor(team): Rename task package to tasks
This commit renames the `task` package to `tasks` within the `ui/team` directory.

The following changes were made:
- Renamed the directory `app/src/main/java/org/ole/planet/myplanet/ui/team/task` to `app/src/main/java/org/ole/planet/myplanet/ui/team/tasks`.
- Updated the package declarations in `TeamTaskAdapter.kt` and `TeamTaskFragment.kt`.
- Updated the import statement in `TeamPageConfig.kt` to reference the new package.
