+++ app/src/main/java/org/ole/planet/myplanet/ui/team/tasks/TeamTaskAdapter.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.team.task
+package org.ole.planet.myplanet.ui.team.tasks
 
 import android.app.AlertDialog
 import android.content.Context
@@ -17,7 +17,7 @@ import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.databinding.RowTaskBinding
 import org.ole.planet.myplanet.model.RealmTeamTask
 import org.ole.planet.myplanet.repository.UserRepository
-import org.ole.planet.myplanet.ui.team.task.TeamTaskAdapter.ViewHolderTask
+import org.ole.planet.myplanet.ui.team.tasks.TeamTaskAdapter.ViewHolderTask
 import org.ole.planet.myplanet.utilities.DiffUtils
 import org.ole.planet.myplanet.utilities.TimeUtils.formatDate
